### 2024-03-26 #6545 @editorialbot
[PRE REVIEW]: ComFiT: a python library for computational field theory with topological defects
**Submitting author:** @vidarsko (<a href="http://orcid.org/0000-0003-4941-6886">Vidar Skogvoll</a>)
**Repository:** https://github.com/vidarsko/ComFiT
**Branch with paper.md** (empty if default branch):
**Version:** v1.4.2
