### 2024-03-26 fb505f7
Update nematic_liquid_crystal.py
