### 2024-03-26 e05a315
Update test_nematic_liquid_crystal.py
