### 2024-03-26 81dc936
Update test_nematic_liquid_crystal.py
