### 2024-03-26 4aea3d9
Update nematic_liquid_crystal.py
