### 2024-03-26 65b4d2f
Removed topological defect documentation
