### 2024-03-26 f57517b
Images to tracking algorithm
