### 2024-03-26 aa537f8
Fixed image typo
