### 2024-03-26 91c88ae
Fixed defect tracking issue near boundary
