+++ docs/TopologicalDefects.md
@@ -1,31 +1,21 @@
 # Topological defects
 
 Topological defects are defects that appear due to imposed constraints by boundary conditions[^merminTopologicalTheoryDefects1979].
-
-In this work, we will track topological defects as course entities derivable from the primary field of observation[^skogvollUnifiedFieldTheory2023].
-
-## Topological defects in a Bose Einstein Condensate
-
-In order to get a sense of what topological defects are, we will look at how they emerge in a Bose-Einstein condensate.
-As simulated in the class BoseEinsteinCondensate, the BEC is represented by a complex function $\psi$, which has a specific modulus $\psi_0$ in equilibrium.
-
-The equilibrium condition of $|\psi|=\psi_0$ does not, however, determine the phase $\theta \equiv arg(\psi)$ of $\psi$.
-Thus the all states with psi0 are equilibrium states.
-The possible equilibrium values of $\theta$ define the _ground state manifold_ $\mathcal M$.
-Since the phase of a complex function is an angle between $0$ and $2\pi$, the ground state manifold is the circle.
-On top of the ground state manifold, one may define the 
+In this work, we will track topological defects as entities derivable from a coarse defect density field[^skogvollUnifiedFieldTheory2023].
+In short, for a system containing topological defects, it is possible to derive a defect density field $\rho$, which upon suitable spatial integration gives the charge of the defects contained in that region.
 
 ## The algorithm for tracking topological defects
 
-The algorithm for tracking topological defects is bases on the coarse-grained description of topological defects presented in Ref.[^skogvollUnifiedFieldTheory2023].
-The algorithm is implemented in the method `BaseSystem.calc_defect_nodes`.
+The algorithm for identifying topological defects is implemented in the method `BaseSystem.calc_defect_nodes`.
 It requires an input of a positive real scalar field `defect_density`, which upon integration over a region gives a number proportional to the number of defects in that region.
-The algorithm takes
 
-* `charge_tolerance` 
-* `integration_radius`
+!!! note "Note"
+    The `defect_density` field is not necessarily the same as the defect density field $\rho$ alluded to above.
+    Whereas $\rho$ might be a vector- or tensor-valued field, `defect_density` is a real and positive scalar field.
+    They are, however, related, and the exact connection depends on the model in question. 
+    For example, in the case of a 2D Bose-Einstein condensate, the `defect_density` is given by $|\rho|$, and in three dimension, where $\vec \rho$ is a vector field, `defect_density` is given by $\vec \rho$. 
 
-as inputs, and then follows these steps:
+The algorithm takes`charge_tolerance` and `integration_radius` as inputs, and then follows these steps:
 
 | Step | Illustration|
 |----|----|
@@ -39,8 +29,26 @@ as inputs, and then follows these steps:
 | &nbsp;&nbsp; 3.5 Identify the `position_index` corresponding to `max(defect_density)` in the region `region_to_search`. | ![Illustration of step 3.5](img/topological_defects_algorithm_3_5.png#only-light) ![Illustration of step 3.5](img/topological_defects_algorithm_3_5-colorinverted.png#only-dark) |
 | &nbsp;&nbsp; 3.6 Calculate the integral `charge` of `defect_density` in a ball `region_to_integrate` of radius `integration_radius` around the point corresponding to `position_index`. | ![Illustration of step 3.6](img/topological_defects_algorithm_3_6.png#only-light) ![Illustration of step 3.6](img/topological_defects_algorithm_3_6-colorinverted.png#only-dark)|
 
-The value of `charge_tolerance` and `integration_radius` must be set depending on the particular system and nature of defect density. 
-For instance,
[…]
