### 2024-03-26 fafa4f9
Fixed BEC test
