+++ comfit/models/nematic_liquid_crystal.py
@@ -135,7 +135,7 @@ class NematicLiquidCrystal(BaseSystem):
             Exception: If the dimension of the system is not 2.
         """
         if not (self.dim == 2):
-            raise Exception("The dimension of the system must be 2 for a vortex dipole configuration.")
+            raise Exception("The dimension of the system must be 2 for a disclination dipole configuration.")
 
         if self.Q is None:
             self.conf_initial_condition_ordered(noise_strength=0)
@@ -574,17 +574,17 @@ class NematicLiquidCrystal(BaseSystem):
         self.Q = np.real(self.Q)
 
 
-##### defect tracking
+##### disclination tracking
     def calc_disclination_density_nematic(self):
         #TODO: See if this can be optimised
         """
-        Calculates the defect density for the nematic. Note that in three dimension the defect density is a tensor
+        Calculates the disclination density for the nematic. Note that in three dimension the disclination density is a tensor
 
         Input:
             None
 
         Output:
-            (numpy.narray) The defect density
+            (numpy.narray) The disclination density
         """
         if self.dim == 2:
             psi0 = np.sqrt(self.B)/2
@@ -686,10 +686,10 @@ class NematicLiquidCrystal(BaseSystem):
 
             return S, n
 
-    def calc_vortex_velocity_field(self, dt_Q):
+    def calc_disclination_velocity_field(self, dt_Q):
         # TODO make 3D
         """
-        Calculates the velocity field of the defect in two dimensions
+        Calculates the velocity field of the disclination in two dimensions
         
         Input:
             dt_Q (numpy.narray) the time derivative of the order parameter
@@ -708,94 +708,94 @@ class NematicLiquidCrystal(BaseSystem):
                                                    [np.real(dt_psi), np.imag(dt_psi)])
 
 
-    def calc_defect_polarization_field(self):
+    def calc_disclination_polarization_field(self):
         ex = np.real(sp.fft.ifftn(1j*self.k[0]*self.Q_f[0] + 1j*self.k[1]*self.Q_f[1]))
         ey = np.real(sp.fft.ifftn(1j * self.k[0] * self.Q_f[1] - 1j * self.k[1] * self.Q_f[0]))
         return np.array([ex,ey])
 
     def calc_disclination_nodes_nem(self, dt_Q=None,polarization = None,charge_tolerance=None):
         """
-        Calculate the positions and charges of vortex nodes based on the defect density.
+        Calculate the positions and charges of disclination nodes based on the disclination density.
         
         Input:
-            dt_Q (numpy.narray, optional): The time derivative of the order parameter. If not provided, the velocity of the vortex nodes will not be calculated.
+            dt_Q (numpy.narray, optional): The time derivative of the order parameter. If not provided, the velocity of the disclination nodes will not be calculated.
         
         Output:
-            list: A list of dictionaries representing the vortex nodes. Each dictionary contains the following keys:
-                  - 'position_index': The position index of the vortex node in the defect density array.
-                  - 'charge': The charge of the vortex node.
-                  - 'position': The position of the vortex node as a list [x, y].
-                  - 'velocity': The velocity of the vortex node as a list [vx, vy].
+            list: A list of dictionaries representing the disclination nodes. Each dictionary contains the following keys:
+                  - 'position_index': The position index of the disclination node in the disclination density array.
+                  - 'charge': The charge of the disclination node.
+                  - 'position': The position of the disclination node as a list [x, y].
+                  - 'velocity': The velocity of the disclination node as a list [vx, vy].
         """
 
-        # Calculate defect density
+        # Calculate disclination density
         if self.dim == 2:
             psi = self.Q[0] + 1j*self.Q[1]
             rho = self.calc_disclination_density_nematic()
 
[…]
