### 2024-03-28 4c37ada
updating tutorial to new notation
