### 2024-04-02 #6545 @sharanroongta
Dear @lucydot . Thanks for the opportunity, but I would pass this one. I have no or very limited working knowledge of QM, so probably someone else is better suited. Thanks.
