### 2024-04-02 #6545 @vidarsko
Thanks for considering!

I want to add to the discussion that the project is mostly about solving physics PDEs in general and less about QM in particular (the QM part mostly serves as an example system for tutorial purposes).

This to not scare off any potential reviewers.
