### 2024-04-15 #6599 @vidarsko
@lucydot I realize that I have not capitalized the p in "Python" for the submission on the JOSS website. Is it possible to amend that?
