### 2024-04-15 #6599 @sharanroongta
@vidarsko Thanks for adding the state of the art. Is it also possible to add references for other open-source library for solving PDE?
