### 2024-04-15 #6599 @sbacchio
@vidarsko, Thank you for promptly addressing the missing item. The added paragraphs are satisfactory to me.
I warmly recommend the work for publication! It is a nice package with a very good documentation and examples.

I have only a marginal suggestion: you should add the example you provided in the manuscript to the tests and CI, such that you guarantee it will always be supported in the future :)
