### 2024-04-18 #6599 @lucydot
> I realize that I have not capitalized the p in "Python" for the submission on the JOSS website. Is it possible to amend that?

Do you mean in the issue thread title @vidarsko? I have just updated.
