### 2024-04-18 #6599 @vidarsko
> > I realize that I have not capitalized the p in "Python" for the submission on the JOSS website. Is it possible to amend that?
>
> Do you mean in the issue thread title @vidarsko? I have just updated.

Thanks for updating. I suppose the p in the issue thread is not capitalized because I forgot to do it in the submission on the JOSS website. This is how my pending submission looks like on the website.

![image](https://github.com/openjournals/joss-reviews/assets/8513212/f1eab84f-78b5-49e9-989b-a9fcc19d819e)
