### 2024-04-20 vidarsko/ComFiT#9 @sharanroongta
documentation
@vidarsko Just wanted to point out the inconsistency in the documentation with respect to docstrings.
For example, `BoseEinsteinCondensate ` has a docstring, but `PhaseFieldCrystal` does not have one.

Also, the example mentioned in the `BoseEinsteinCondensate` docstring gives an error:
```python
import comfit
bec = comfit.BoseEinsteinCondensate(3, 100, dissipation_factor=0.5)
```
error:
`TypeError: __init__() takes 2 positional arguments but 3 were given`

Basically `x_resolution` does not exist
