### 2024-04-20 #6599 @sbacchio
> Thanks for your timely review @sbacchio ✨ - can you confirm that as part of the review you were able to install and run the package successfully? I ask as it is unusual for there to be no hiccups, so want to double check..

@lucydot, sure I've been able to install it on my machine and reproduce examples and tutorials. At least with my Python version, it worked out fine!
