### 2024-04-22 9f7361c
Update ClassBoseEinsteinCondensate.md
