### 2024-04-22 4224221
Update ClassNematicLiquidCrystal.md
