### 2024-04-22 059b15c
Update ClassNematicLiquidCrystal.md
