+++ docs/ClassNematicLiquidCrystal.md
@@ -27,10 +27,18 @@ The NematicLiquidCrystal class takes the same keyword as the BaseSystem class in
 
 | Keyword | Definition | Default value|
 |---------|------------|--------------|
-| `gamma`  | dissipative factor | $0.01$ |
+| `alpha`  | The activity parameter. See model section for sign convention | $-1$ |
+| `K`  | Frank elastic constant  | $1$ |
+| `A`  | Parameter in front of $\text{Tr}(Q^2)$ in the free-energy | $1$ |
+| `B`  | Parameter in front of $\text{Tr}(Q^2)^2$ in the free-energy | $1$ |
+| `C`  | Parameter in front of $\text{Tr}(Q^3)$ in the free-energy. 3D only | $0$ |
+| `gamma`  | Rotational diffusion | $1$ |
+| `Gamma`  | Friction  | $0$ |
+| `eta`  | Viscosity | $1$ |
 
+These parameters are discussed in more detail in the model section. 
 
-### The parameters
+### Note on the tensor parameters
 The $Q$ tensor is given by the nematic director as
 
 $$
