### 2024-04-22 1fef0c1
Update ClassNematicLiquidCrystal.md
