### 2024-04-25 2028fd0
replaced input with args
