+++ comfit/models/nematic_liquid_crystal.py
@@ -18,7 +18,7 @@ class NematicLiquidCrystal(BaseSystem):
         """
         Initializes a system to simulate a (active) nematic liquid crystal
 
-        Input:
+        Args:
         - dimension : int
             The dimension of the system.
         - kwargs : dict, optional
@@ -70,7 +70,7 @@ class NematicLiquidCrystal(BaseSystem):
         Initialises the system with the nematogens pointing in the x-direction in 2D and in the z-direction in 3D
         with some random noise in the angle.
         
-        Input:
+        Args:
              noise_strength (float): A meshure for how much noise to put in the angle
         
         Output:
@@ -125,7 +125,7 @@ class NematicLiquidCrystal(BaseSystem):
         """
         Sets the initial condition for a disclination dipole configuration in a 2-dimensional system.
 
-        Input:
+        Args:
             None
 
         Output:
@@ -156,7 +156,7 @@ class NematicLiquidCrystal(BaseSystem):
         Sets the initial condition for a disclination line in a 3-dimensional system.
         The dislocation is parralell to the z-axis
 
-        Input:
+        Args:
             position (list): the position of the dislocation. Only the position in the xy plane is used
             angle (float): the angle between the director and the z axis.
             sign (float): +1 or -1. Charge of the dislocation in the xy-plain
@@ -202,7 +202,7 @@ class NematicLiquidCrystal(BaseSystem):
         """
         Set the activity to zero everywhere exept for inside a channel of width "width"
         
-        Input:
+        Args:
             width (float): width of the channel
             d (float, optional): width of interface
         
@@ -222,7 +222,7 @@ class NematicLiquidCrystal(BaseSystem):
         '''
         Updates the velocity and its fourier transform given a nematic field Q.
         
-        Input:
+        Args:
             (numpy.narray) the Q tensor
         
         Output:
@@ -240,7 +240,7 @@ class NematicLiquidCrystal(BaseSystem):
         '''
         Function that calculates the activ force in Fourier space.
         
-        Input:
+        Args:
             Q (numpy.narray) the order parameter that we use to find the force.
         
         Output:
@@ -255,7 +255,7 @@ class NematicLiquidCrystal(BaseSystem):
         '''
         Calculate the passive force in Fourier space
         
-        Input:
+        Args:
             Q (numpy.narray) the order parameter that we use to find the force.
         
         Output: 
@@ -271,7 +271,7 @@ class NematicLiquidCrystal(BaseSystem):
         """
         Calculates the passive stress in fourier space
 
-        Input:
+        Args:
             Q (numpy.narray) the order parameter that we use to find the stress.
         
         Output: 
@@ -343,7 +343,7 @@ class NematicLiquidCrystal(BaseSystem):
         """
         Finds the molecular field (NB! need to be rewriten when C != 0)
         
-        Input:
+        Args:
             Q (numpy.ndarray): The nematic tensor
         
         Output:
@@ -372,7 +372,7 @@ class NematicLiquidCrystal(BaseSystem):
         '''
         Calculates the pressure in Fourier space. The zero mode is set to zero
         
-        Input:
+        Args:
             F_af (numpy.narray) the active force in Fourier space
             F_pf (numpy.narray) the passive force in Fourier space
         
@@ -388,7 +388,7 @@ class NematicLiquidCrystal(BaseSystem):
         """
         Caclulates the gradient of the pressure
         
-        Input: 
+        Args: 
             p_f (numpy.narray) the pressure in Fourier space
         
         Output:
@@ -403,7 +403,7 @@ class NematicLiquidCrystal(BaseSystem):
         """
         Calculates the vorticity tensor
 
-        Input:
+        Args:
             None
 
         Output:
@@ -429,7 +429,7 @@ class NematicLiquidCrystal(BaseSystem):
[…]
