+++ comfit/core/base_system_calc.py
@@ -10,15 +10,14 @@ class BaseSystemCalc:
     def calc_angle_field_single_vortex(self,
                                        position=None,
                                        charge=1):
-        """
-        Calculate the angle field due to a single vortex.
+        """Calculate the angle field due to a single vortex.
 
         Args:
             position (list, optional): The position of the vortex. Defaults to None.
             charge (int, optional): The charge of the vortex. Defaults to 1.
         
         Returns:
-            numpy.ndarray: The angle field calculated for the vortex.
+            The angle field calculated for the vortex. numpy.ndarray.
             
         Raises:
             Exception: If the dimension of the system is not 2.
@@ -39,14 +38,14 @@ class BaseSystemCalc:
         """Calculates the angle field for a double vortex system.
 
         Args:
-            dipole_vector (list, optional): The dipole vector. Defaults to None.
-            dipole_position (list, optional): The position the center of mass of the dipole. Defaults to None.
+            dipole_vector: The dipole vector. List or tuple. 
+            dipole_position: The position the center of mass of the dipole. 
+            
+        Returns:
+            The calculated angle field for the double vortex system as a numpy array. numpy.ndarray.
 
         Raises:
             Exception: If the dimension of the system is not 2.
-            
-        Returns:
-            np.ndarray: The calculated angle field for the double vortex system.
         """
 
         if self.dim != 2:
@@ -86,14 +85,15 @@ class BaseSystemCalc:
         return np.mod(theta + np.pi, 2 * np.pi) - np.pi
 
     def calc_angle_field_vortex_ring(self, position=None, radius=None, normal_vector=[0, 0, 1]):
-        """
-        Calculates the angle field for a vortex ring.
+        """Calculates the angle field for a vortex ring.
+
         Args:
-            position (list, optional): The position of the vortex ring. Defaults to None.
-            radius (float, optional): The radius of the vortex ring. Defaults to None.
-            normal_vector (list, optional): The normal vector of the vortex ring. Defaults to [0,0,1].
+            position: The position of the vortex ring. 
+            radius: The radius of the vortex ring. 
+            normal_vector: The normal vector of the vortex ring.
+
         Returns:
-            numpy.ndarray: The calculated angle field for the vortex ring.
+            The calculated angle field for the vortex ring.
         """
         if position is None:
             position = self.rmid
@@ -155,21 +155,20 @@ class BaseSystemCalc:
         return np.mod(theta + np.pi, 2 * np.pi) - np.pi
 
     def calc_wavenums(self, x):
-        """
-        Calculates the wavenumbers corresponding to the input position vectors given by x.
+        """Calculates the wavenumbers corresponding to the input position vectors given by x.
 
         Args:
-        - x : 1D array of x-positions.
+            x: 1D array of x-positions.
 
         Returns:
-            numpy array: 1D array of wavenumbers with all the modes for the given x-array,
-            assuming periodicity from x[0] to x[0] over n intervals.
+            1D array of wavenumbers with all the modes for the given x-array,
+            assuming periodicity from x[0] to x[0] over n intervals. numpy array.
 
         Example:
-        x = np.array([-10, -5, 0, 5, 10])
-        k = instance_of_BaseSystem.calc_wavenums(self,x)
-        print(k)
-        # Returns: [ 0.          0.25132741  0.50265482 -0.50265482 -0.25132741]
+            x = np.array([-10, -5, 0, 5, 10])
+            k = instance_of_BaseSystem.calc_wavenums(self,x)
+            print(k)
+            # Returns: [ 0.          0.25132741  0.50265482 -0.50265482 -0.25132741]
         """
         n = len(x)
 
[…]
