+++ comfit/core/base_system_conf.py
@@ -2,4 +2,5 @@ import numpy as np
 import scipy as sp
 
 class BaseSystemConf:
+    """ Configuration methods for the base system class"""
     pass
