### 2024-04-25 1972285
Fixed doc in base_system_get
