+++ comfit/core/base_system_get.py
@@ -2,19 +2,17 @@ import numpy as np
 import scipy as sp
 
 class BaseSystemGet:
-    # GET FUNCTIONS
-
+    """ Get methods for the base system class"""
     def get_sym(self,tensor,i,j):
-        """
-        Gets the i,j component of a symmetric tensor saved in an array structure.
+        """Gets the i,j component of a symmetric tensor saved in an array structure.
 
         Args: 
             tensor (numpy.ndarray): The symmetric tensor.
-            i (int): The row index.
-            j (int): The column index.
+            i: The row index. (int)
+            j: The column index. (int)
         
         Returns:
-            (numpy.ndarray) The i,j component of the tensor.
+            The i,j component of the tensor. (numpy.ndarray) 
         """
 
         if self.dim == 2:
@@ -32,16 +30,15 @@ class BaseSystemGet:
                 return tensor[2] if j == 0 else tensor[4] if j == 1 else tensor[5]
 
     def get_sym_tl(self,tensor,i,j):
-        """
-        Gets the i,j component of a symmetric traceless tensor saved in an array structure.
+        """Gets the i,j component of a symmetric traceless tensor saved in an array structure.
 
         Args:
-            tensor (numpy.ndarray): The symmetric traceless tensor.
-            i (int): The row index.
-            j (int): The column index.
+            tensor: The symmetric traceless tensor. (numpy.ndarray)
+            i: The row index. (int)
+            j: The column index. (int)
         
         Returns:
-            (numpy.ndarray) The i,j component of the tensor.
+            The i,j component of the tensor. (numpy.ndarray) 
         """
         if self.dim == 2:
             if i == 0:
@@ -58,16 +55,15 @@ class BaseSystemGet:
                 return tensor[2] if j == 0 else tensor[4] if j == 1 else -(tensor[0] + tensor[3])
 
     def get_anti_sym(self,omega,i,j):
-        """
-        Gets the i,j component of an anti-symmetric tensor saved in an array structure.
+        """Gets the i,j component of an anti-symmetric tensor saved in an array structure.
 
         Args:
-            omega (numpy.ndarray): The anti-symmetric tensor.
-            i (int): The row index.
-            j (int): The column index.
+            omega: The anti-symmetric tensor. (numpy.ndarray)
+            i: The row index. (int)
+            j: The column index. (int)
         
         Returns:
-            (numpy.ndarray) The i,j component of the tensor.
+            The i,j component of the tensor. (numpy.ndarray) 
         """
         # TODO: I don't like that the input vector is a scalar field in 2 dimensions. (Vidar 11.03.24)
         if self.dim == 2:
