### 2024-04-25 54d8e92
Fixed doc in base_system_init
