### 2024-04-25 d8cbe19
Update base_system_plot.py
