+++ comfit/core/base_system_plot.py
@@ -10,14 +10,16 @@ import matplotlib.colors as mcolors
 import matplotlib.tri as mtri
 
 class BaseSystemPlot:
-    # PLOTTING FUNCTIONS
+    """ Plotting methods for the base system class"""
 
     def plot_tool_set_axis_properties(self, **kwargs):
-        """
-        Sets the properties of the axis for a plot.
+        """Sets the properties of the axis for a plot.
         
         Args:
             kwargs: keyword arguments for the axis properties
+
+        Returns:
+            The axis object with the properties set.
         """
 
         # Check if an axis object is provided
@@ -154,13 +156,15 @@ class BaseSystemPlot:
         return ax
 
     def plot_tool_extend_field(self,field):
-        """
-        Extends the field in case not a complete array is given. 
+        """Extends the field in case not a complete array is given. 
+
         For instance, if a field in 3 dimensions is calculated using only x, then the field is extended to the full 3D array.
+
         Args:
             field: The field to be extended
+
         Returns:
-            np.ndarray: The extended field
+            The extended field (np.ndarray)
         """    
         # 2 dimensional fields
         if field.shape == (self.xRes,1):
@@ -194,8 +198,13 @@ class BaseSystemPlot:
         return field
 
     def plot_tool_surface(self, **kwargs):
-        """
-        Plots the surface of the given field.
+        """Plots the surface of the given field.
+
+        Args:
+            **kwargs: Keyword arguments for the plot.
+        
+        Returns:
+            The axes containing the plot.
         """
         field = kwargs['field']
         value = kwargs['value']
@@ -210,8 +219,7 @@ class BaseSystemPlot:
         ax.plot_trisurf(x, y, faces, z, alpha=alpha, color=color)
     
     def plot_field(self, field, **kwargs):
-        """
-        Plots the given (real) field.
+        """Plots the given (real) field.
         
         Args:
             field (array-like): The field to be plotted.
@@ -220,7 +228,7 @@ class BaseSystemPlot:
                 for a full list of keyword arguments.
         
         Returns:
-            matplotlib.axes.Axes: The axes containing the plot.
+            The axes containing the plot (matplotlib.axes.Axes).
         """
 
         if field.dtype == bool:
@@ -427,7 +435,6 @@ class BaseSystemPlot:
         
         Returns:
             matplotlib.axes.Axes: The axes containing the plot.
-
         """
 
         # Check if an axis object is provided
@@ -651,15 +658,14 @@ class BaseSystemPlot:
         return fig, ax
 
     def plot_angle_field(self, angle_field, **kwargs):
-        """
-        Plot the angle field.
+        """Plot the angle field.
 
         Args:
             field (array-like): The angle field values.
             ax (matplotlib.axes.Axes, optional): The axes to plot the angle field on. If not provided, a new subplot will be created.
         
         Returns:
-            matplotlib.axes.Axes: The axes containing the plot.
+            The axes containing the plot. (matplotlib.axes.Axes)
         """
 
         # Check if the vector field is complex
@@ -694,17 +700,17 @@ class BaseSystemPlot:
             return self.plot_complex_field(complex_field, **kwargs)
 
     def plot_vector_field(self, vector_field, spacing=5, **kwargs):
-        """
-        Plots a vector field.
+        """Plots a vector field.
 
         Args:
-        vector_field (tuple): Tuple containing the x and y components of the vector field.
-        spacing (int, optional): The spacing for the quiver plot. Default is 5.
-        kwargs: Keyword arguments for the plot, see https://vidarsko.github.io/ComFiT/Plotting/.
[…]
