### 2024-04-25 398b428
Fixed doc in base_system_plot
