### 2024-04-25 c4d62f2
Fixed doc in bose_einstein_condensate
