+++ comfit/models/bose_einstein_condensate.py
@@ -7,24 +7,22 @@ import scipy as sp
 
 class BoseEinsteinCondensate(BaseSystem):
     def __init__(self, dimension, **kwargs):
-        """
-        Initializes a system to simulate a Bose-Einstein Condensate using the Gross-Pitaevskii equation.
+        """Initializes a system to simulate a Bose-Einstein Condensate using the Gross-Pitaevskii equation.
 
         Args:
-        - dimension : int
-            The dimension of the system.
-        - kwargs : dict, optional
-            Optional keyword arguments to set additional parameters. see
-            https://vidarsko.github.io/ComFiT/ClassBoseEinsteinCondensate/
+            dimension : int
+                The dimension of the system.
+            kwargs : dict, optional
+                Optional keyword arguments to set additional parameters. see
+                https://vidarsko.github.io/ComFiT/ClassBoseEinsteinCondensate/
 
         Returns:
-        - BoseEinsteinCondensate object
-            The system object representing the BoseEinsteinCondensate simulation.
+            The system object representing the BoseEinsteinCondensate simulation. BoseEinsteinCondensate object
 
         Example:
-        bec = BoseEinsteinCondensate(3,xRes=101,yRes=101,zRes=101, gamma=0.5)
-        Creates a BoseEinsteinCondensate system with 3 dimensions and a spatial resolution of 101 in all directions.
-        The dissipative factor gamma is set to 0.5.
+            bec = BoseEinsteinCondensate(3,xRes=101,yRes=101,zRes=101, gamma=0.5)
+            Creates a BoseEinsteinCondensate system with 3 dimensions and a spatial resolution of 101 in all directions.
+            The dissipative factor gamma is set to 0.5.
         """
 
         # First initialize the BaseSystem class
@@ -45,16 +43,20 @@ class BoseEinsteinCondensate(BaseSystem):
             setattr(self, key, value)
 
     def __str__(self):
-        """
-        Output a string representation of the system.
+        """Output a string representation of the system.
+
+        Args:
+            None
+
+        Returns:
+            A string representation of the system.
         """
         return f"ComFiT object: BoseEinsteinCondensate\n \
                 Dimension: {self.dim}\n"
 
     # CONFIGURATION FUNCTIONS
     def conf_initial_condition_disordered(self, noise_strength=0.01):
-        """
-        Sets disordered initial condition for the BoseEinsteinCondensate with some thermal flcutiations
+        """Sets disordered initial condition for the BoseEinsteinCondensate with some thermal flcutiations
         
         Args:
             noise_strength: the strength of the noise
@@ -81,11 +83,12 @@ class BoseEinsteinCondensate(BaseSystem):
         self.psi_f = sp.fft.fftn(self.psi)
 
     def conf_external_potential(self, V_ext, additive=False):
-        """
-        Sets the external potential of the system.
+        """Sets the external potential of the system.
+
         Args:
             V_ext (function or float): the external potential
             additive (bool, optional): whether to add the new potential to the existing potential or not
+
         Returns:
             Modifies the value of self.V_ext
         """
@@ -102,15 +105,15 @@ class BoseEinsteinCondensate(BaseSystem):
 
 
     def conf_initial_condition_Thomas_Fermi(self):
-        """
-        Finds the Thomas_Fermi ground state.
+        """Finds the Thomas_Fermi ground state.
+
         Must be precided by an energy relaxation to find the true ground state
 
         Args:
             None
         
         Returns: 
-            sets the value of self.psi and self.psi_f
+            Sets the value of self.psi and self.psi_f
         """
         V_0 = np.zeros(self.dims) + self.V_ext(self.time)
         self.psi = np.emath.sqrt(1 - V_0)
@@ -120,8 +123,7 @@ class BoseEinsteinCondensate(BaseSystem):
 
[…]
