+++ comfit/core/base_system_init.py
@@ -10,7 +10,7 @@ class BaseSystemInit:
         Args:
             dimension (int): The dimension of the system. Must be 1, 2, or 3.
             **kwargs: Additional keyword arguments, see 
-            https://vidarsko.github.io/ComFiT/ClassBaseSystem/
+            https://comfitlib.com/ClassBaseSystem/
 
         Returns:
             None
