### 2024-04-25 7e82203
Fixed doc in nematic_liquid_crystal
