+++ comfit/models/bose_einstein_condensate.py
@@ -613,7 +613,7 @@ class BoseEinsteinCondensate(BaseSystem):
                                  - 'position': The position of the vortex node as a list [x, y].
                                  - 'velocity': The velocity of the vortex node as a list [vx, vy].
             -**kwargs: Keyword arguments for the plot.
-                See github.com/vidarsko/ComFiT/blob/main/docs/ClassBaseSystem.md
+                See https://comfitlib.com/ClassBaseSystem/
                 for a full list of keyword arguments.
         
         Returns:
