+++ comfit/models/phase_field_crystal.py
@@ -696,7 +696,7 @@ class PhaseFieldCrystal(BaseSystem):
         Args:
             - dislocation_nodes: The dislocation nodes to plot.
             -**kwargs: Keyword arguments for the plot.
-                See github.com/vidarsko/ComFiT/blob/main/docs/ClassBaseSystem.md 
+                See https://comfitlib.com/ClassBaseSystem/ 
                 for a full list of keyword arguments.  
 
         Returns: 
