+++ comfit/models/nematic_liquid_crystal.py
@@ -28,9 +28,9 @@ class NematicLiquidCrystal(BaseSystem):
             The system object representing the nematic simulation.
 
         Example:
-        nematic = NematicLiquidCrystal(2, xRes=100, yRes = 100, alpha=-0.5)
-        Creates a nematic liquid crystal with 2 dimensions and a spatial resolution of 100.
-        The activity alpha is set to -0.5.
+            nematic = NematicLiquidCrystal(2, xRes=100, yRes = 100, alpha=-0.5)
+            Creates a nematic liquid crystal with 2 dimensions and a spatial resolution of 100.
+            The activity alpha is set to -0.5.
         """
         super().__init__(dimension, **kwargs)
 
@@ -57,22 +57,26 @@ class NematicLiquidCrystal(BaseSystem):
             setattr(self, key, value)
 
     def __str__(self):
-        """
-        Output a string representation of the system.
+        """Output a string representation of the system.
+
+        Input:
+            None
+
+        Returns:
+            A string representation of the system.
         """
         return "NematicLiquidCrystal"
 
     # Initial condition
     def conf_initial_condition_ordered(self, noise_strength=0.01):
-        """
-        Initialises the system with the nematogens pointing in the x-direction in 2D and in the z-direction in 3D
+        """Configures the system with the nematogens pointing in the x-direction in 2D and in the z-direction in 3D
         with some random noise in the angle.
         
         Args:
-             noise_strength (float): A meshure for how much noise to put in the angle
+             noise_strength: A meshure for how much noise to put in the angle (float)
         
         Returns:
-            Initialises self.Q  and self.Q_f
+            Initialises self.Q and self.Q_f
         
         Raises:
             Exception if the dimension is not 2 or 3
@@ -120,14 +124,13 @@ class NematicLiquidCrystal(BaseSystem):
             raise Exception("This dimension is not included for the moment")
 
     def conf_insert_disclination_dipole(self, dipole_vector=None, dipole_position=None):
-        """
-        Sets the initial condition for a disclination dipole configuration in a 2-dimensional system.
+        """Sets the initial condition for a disclination dipole configuration in a 2-dimensional system.
 
         Args:
             None
 
         Returns:
-            None
+            Configures self.Q and self.Q_f with a disclination dipole configuration.
 
         Raises:
             Exception: If the dimension of the system is not 2.
@@ -150,8 +153,8 @@ class NematicLiquidCrystal(BaseSystem):
         self.Q_f = sp.fft.fft2(self.Q)
 
     def conf_initial_disclination_lines(self, position=None):
-        """
-        Sets the initial condition for a disclination line in a 3-dimensional system.
+        """Sets the initial condition for a disclination line in a 3-dimensional system.
+
         The dislocation is parralell to the z-axis
 
         Args:
@@ -197,15 +200,14 @@ class NematicLiquidCrystal(BaseSystem):
 
 
     def conf_active_channel(self,width,d=7):
-        """
-        Set the activity to zero everywhere exept for inside a channel of width "width"
+        """Configures the activity to zero everywhere exept for inside a channel of width "width"
         
         Args:
-            width (float): width of the channel
-            d (float, optional): width of interface
+            width: width of the channel (float)
+            d: width of interface (float, optional)
         
         Returns:
-            updates the activity to the channel configuration.
+            Updates the activity to the channel configuration.
         """
         if self.dim ==2:
             X, Y = np.meshgrid(self.x, self.y, indexing='ij')
[…]
