### 2024-04-25 e5fa036
Fixed doc in quantum_mechanics
