+++ comfit/models/quantum_mechanics.py
@@ -6,22 +6,18 @@ import scipy as sp
 
 class QuantumMechanics(BaseSystem):
     def __init__(self, dimension, **kwargs):
-        """
-        Initializes a quamtum mechanics system evolving according to the Schrödinger equation
+        """Initializes a quamtum mechanics system evolving according to the Schrödinger equation
 
         Args:
-        - dimension : int
-            The dimension of the system.
-        - kwargs : dict, optional
-            Optional keyword arguments to set additional parameters. Same as BaseSystem
-
+            dimension: The dimension of the system.
+            **kwargs : dict, optional. Optional keyword arguments to set additional parameters. Same as BaseSystem
+                
         Returns:
-        - QuantumMechanics object
             The system object representing the QuantumMechanics simulation.
 
         Example:
-        qm = cf.QuantumMechanics(3,xRes=101,yRes=101,zRes=101)
-        Creates a BoseEinsteinCondensate system with 3 dimensions and a spatial resolution of 101.
+            qm = cf.QuantumMechanics(3,xRes=101,yRes=101,zRes=101)
+            Creates a BoseEinsteinCondensate system with 3 dimensions and a spatial resolution of 101.
         """
 
         # First initialize the BaseSystem class
@@ -40,14 +36,26 @@ class QuantumMechanics(BaseSystem):
             setattr(self, key, value)
 
     def __str__(self):
-        """
-        Output a string representation of the system
+        """Output a string representation of the system
+
+        Args:
+            None
+
+        Returns:
+            A string representation of the system
         """
         return 'QuantumMechanics'
 
     def conf_initial_condition_Gaussian(self,position=None,width=None, initial_velocity=None):
-        """
-        Set the initial condition to a Gaussian wavepacket.
+        """Set the initial condition to a Gaussian wavepacket.
+
+        Args:
+            position: The position of the wavepacket.
+            width: The width of the wavepacket.
+            initial_velocity: The initial velocity of the wavepacket.
+
+        Returns:
+            Configures self.psi and self.psi_f.
         """
 
         if position == None:
@@ -75,10 +83,13 @@ class QuantumMechanics(BaseSystem):
 
 
     def conf_harmonic_potential(self,trapping_strength=None):
-        """
-        Set the external potential to a harmonic trap with R_tf being the thomas fermi radius
-        :param R_tf:
-        :Returns:
+        """Set the external potential to a harmonic trap with R_tf being the thomas fermi radius
+        
+        Args:
+            trapping_strength: The strength of the trapping potential.
+
+        Returns:
+            Configures self.V_ext.
         """
 
         if trapping_strength == None:
@@ -96,16 +107,7 @@ class QuantumMechanics(BaseSystem):
                                            +((self.z - self.zmid) ** 2) )
 
     def conf_hydrogen_state(self,n,l,m):
-        """
-        Set the initial condition to a hydrogen state with quantum numbers n,l,m
-        """
-
-        self.psi = self.calc_hydrogen_state(n,l,m)
-        self.psi_f = sp.fft.fftn(self.psi)
-
-    def calc_hydrogen_state(self,n,l,m):
-        """
-        Calculate the hydrogen state with quantum numbers n,l,m
+        """Set the initial condition to a hydrogen state with quantum numbers n,l,m
 
         Args:
             n: principal quantum number
[…]
