+++ comfit/models/phase_field_crystal.py
@@ -15,19 +15,11 @@ class PhaseFieldCrystal(BaseSystem):
         phase field crystal models implemented in comfit
 
         Args:
-        - dimension : int
-            The dimension of the system.
-        - kwargs : dict, optional
-            Optional keyword arguments to set additional parameters. See
-            https://comfitlib.com/ClassPhaseFieldCrystal/
+            dimension: The dimension of the system.
+            kwargs :keyword arguments to set additional parameters. See https://comfitlib.com/ClassPhaseFieldCrystal/
 
         Returns:
-        - PhaseFieldCrystal object
             The system object representing the phase field crystal simulation.
-
-        Example:
-        pfc = PhaseFieldCrystal(2, xRes=100, yRes = 100)
-        Creates a phase field crystal system with 2 dimensions and a spatial resolution of 100.
         """
         
         # First initialize the BaseSystem class
@@ -51,6 +43,12 @@ class PhaseFieldCrystal(BaseSystem):
     def __str__(self):
         """
         Returns a string representation of the object.
+
+        Args:
+            None
+        
+        Returns: 
+            A string representation of the object.
         """
         string = "-------------------------------\n"
         string += "Phase Field Crystal instance\n"
@@ -95,6 +93,14 @@ class PhaseFieldCrystal(BaseSystem):
 
     # CONFIGURATION FUNCTIONS
     def conf_PFC_from_amplitudes(self, eta=None):
+        """Configures the PFC from the amplitudes.
+
+        Args:
+            eta: The amplitudes to configure the PFC from.
+
+        Returns:
+            Configures self.psi and self.psi_f.
+        """
 
         self.psi = self.psi0
 
@@ -116,8 +122,13 @@ class PhaseFieldCrystal(BaseSystem):
         self.psi_f = sp.fft.fftn(self.psi)
 
     def conf_advect_PFC(self,u):
-        """
-        Advects the PFC according to the displacement field u.
+        """Advects the PFC according to the displacement field u.
+
+        Args:
+            u: The displacement field to advect the PFC with.
+
+        Returns:
+            The advected PFC.
         """
 
         self.psi = np.real(self.calc_advect_field(self.psi, u, self.psi_f))
@@ -134,6 +145,15 @@ class PhaseFieldCrystal(BaseSystem):
 
     # EVOLUTION FUNCTIONS
     def evolve_PFC(self, number_of_steps, method='ETD2RK'):
+        """Evolves the PFC according to classical PFC dynamics.
+
+        Args:
+            number_of_steps: The number of steps to evolve the PFC.
+            method: The method to use for the evolution (default: ETD2RK).
+        
+        Returns:
+            Updates self.psi and self.psi_f
+        """
 
         omega_f = self.calc_omega_f()
 
@@ -149,13 +169,15 @@ class PhaseFieldCrystal(BaseSystem):
         self.psi_f = sp.fft.fftn(self.psi)
 
     def evolve_PFC_mechanical_equilibrium(self, time, Delta_t = 10, method='ETD2RK'):
-        """
-        Evolves the PFC in mechanical equilibrium. 
+        """Evolves the PFC in mechanical equilibrium. 
 
         Args:
             number_of_steps: The number of steps to evolve the PFC
             Delta_t: The time step for the mechanical equilibrium evolution
-            method: The method to use for the evolution (default: ETD2RK)
+            method: The method to use for the evolution
+
+        Returns:
+            Updates self.psi and self.psi_f
         """
         
         number_of_steps = round(time/self.dt)
@@ -172,8 +194,19 @@ class PhaseFieldCrystal(BaseSystem):
                                 method = 'ETD2RK',
                                 gamma_S = 2**-2,
                                 rho0 = 2**-2):
-        """
[…]
