### 2024-04-25 9924a49
Added doc to pfc 1d periodic
