+++ comfit/models/phase_field_crystal_2d_square.py
@@ -9,8 +9,14 @@ from pprint import pprint
 
 class PhaseFieldCrystal2DSquare(PhaseFieldCrystal):
     def __init__(self, nx, ny, **kwargs):
-        """
-        Nothing here yet
+        """Initializes a phase field crystal system in 2D with a square crystal structure.
+
+        Args:
+            nx: The number of unit cells in the x direction.
+            ny: The number of unit cells in the y direction.
+
+        Returns:
+            The system object representing the PhaseFieldCrystal2DSquare simulation.
         """
 
         # Type of the system
@@ -65,6 +71,15 @@ class PhaseFieldCrystal2DSquare(PhaseFieldCrystal):
 
         
     def calc_proto_amplitudes_conserved(self):
+        """Calculates the proto-amplitudes for the system.
+
+        Args:
+            None
+
+        Returns:
+            The proto-amplitudes for the system.
+        """
+
         A = 0
         B = 0
         free_energy = self.calc_free_energy_from_proto_amplitudes(self.psi0, A, B)
@@ -82,7 +97,14 @@ class PhaseFieldCrystal2DSquare(PhaseFieldCrystal):
         return A, B
 
     def calc_proto_amplitude_equations_conserved(self,vars):
+        """Calculates the equations for the proto-amplitudes for the system in the case of conserved dynamics
 
+        Args:
+            vars: The proto-amplitudes for the system.
+
+        Returns:
+            The equations for the proto-amplitudes for the system.
+        """
         psi0 = self.psi0
         r = self.r
         t = self.t
@@ -95,18 +117,41 @@ class PhaseFieldCrystal2DSquare(PhaseFieldCrystal):
         return [eq1, eq2]
 
     def calc_free_energy_from_proto_amplitudes(self,psi0,A,B):
+        """Calculates the free energy of the system from the proto-amplitudes.
+
+        Args:
+            psi0: The average value of psi.
+            A: The proto-amplitude.
+            B: The proto-amplitude.
+
+        Returns:
+            The free energy of the system.
+        """
         r = self.r
         t = self.t
         v = self.v
         return (108*A**4*v + 108*B**4*v + psi0**2*(24 + 6*r + 4*t*psi0 + 3*v*psi0**2) + 24*B**2*(r + psi0*(2*t + 3*v*psi0)) + 24*A**2*(r + 18*B**2*v + 4*B*(t + 3*v*psi0) + psi0*(2*t*+ 3*v*psi0)))
 
     def calc_omega_f(self):
+        """Calculates the linear time evolution operator in Fourier space.
+
+        Args:
+            None
+
+        Returns:
+            The linear time evolution operator in Fourier space.
+        """
         k2 = self.calc_k2()
         return - k2 * (self.r + (1 - k2)**2*(2-k2)**2)
 
     def calc_stress_tensor_microscopic(self):
-        """
-        Calculates the microscopic stress of the phase-field crystal.
+        """Calculates the microscopic stress of the phase-field crystal.
+
+        Args:
+            None
+
+        Returns:
+            The microscopic stress of the phase-field crystal.
         """
         stress = np.zeros((3,self.xRes,self.yRes))
         
@@ -120,6 +165,14 @@ class PhaseFieldCrystal2DSquare(PhaseFieldCrystal):
         return stress
 
     def calc_stress_divergence_f(self, field_f = None):
+        """Calculates the divergence of the stress tensor in Fourier space.
+        
+        Args:
+            field_f: The field in Fourier space.
+        
+        Returns:
+            The divergence of the stress tensor in Fourier space.
+        """
         
         if field_f is None:
             field_f = self.psi_f
