### 2024-04-25 f77a650
Fixed docs in pfc 2d triangular
