+++ comfit/models/phase_field_crystal_3d_face_centered_cubic.py
@@ -9,8 +9,15 @@ from pprint import pprint
 
 class PhaseFieldCrystal3DFaceCenteredCubic(PhaseFieldCrystal):
     def __init__(self, nx, ny, nz, **kwargs):
-        """
-        Nothing here yet
+        """Initializes a phase field crystal system in 3D with a face centered cubic crystal structure.
+
+        Args:
+            nx: The number of unit cells in the x direction.
+            ny: The number of unit cells in the y direction.
+            nz: The number of unit cells in the z direction.
+
+        Returns:
+            The system object representing the PhaseFieldCrystal3DFaceCenteredCubic simulation.
         """
 
         # Type of the system
@@ -79,6 +86,15 @@ class PhaseFieldCrystal3DFaceCenteredCubic(PhaseFieldCrystal):
 
         
     def calc_proto_amplitudes_conserved(self):
+        """Calculates the proto-amplitudes for the system.
+
+        Args:
+            None
+        
+        Returns:
+            The proto-amplitudes for the system.
+        """
+
         A = 0
         B = 0
         free_energy = self.calc_free_energy_from_proto_amplitudes(self.psi0, A, B)
@@ -96,7 +112,14 @@ class PhaseFieldCrystal3DFaceCenteredCubic(PhaseFieldCrystal):
         return A, B
 
     def calc_proto_amplitude_equations_conserved(self, vars):
+        """Calculates the equations for the proto-amplitudes for the system in the case of conserved dynamics
 
+        Args:
+            vars: The proto-amplitudes for the system.
+        
+        Returns:
+            The equations for the proto-amplitudes for the system in the case of conserved dynamics.
+        """
         psi0 = self.psi0
         r = self.r
         t = self.t
@@ -109,18 +132,43 @@ class PhaseFieldCrystal3DFaceCenteredCubic(PhaseFieldCrystal):
         return [eq1, eq2]
 
     def calc_free_energy_from_proto_amplitudes(self, psi0, A, B):
+        """Calculates the free energy of the system from the proto-amplitudes.
+
+        Args:
+            psi0: The average value of psi.
+            A: The proto-amplitude.
+            B: The proto-amplitude.
+
+        Returns:
+            The free energy of the system.
+        """
+
         r = self.r
         t = self.t
         v = self.v
         return 2*np.pi**3/np.sqrt(3)*(1944*A**4*v + 810*B**4*v + psi0**2*(32 + 18*r + 12*t*psi0 + 9*v*psi0**2) + 108*B**2*(r + psi0*(2*t + 3*v*psi0)) + 144*A**2*( r + 36*B**2*v + 6*B*(t + 3*v*psi0) + psi0*(2*t + 3*v*psi0)))
 
     def calc_omega_f(self):
+        """Calculates the linear evolution operator of the system.
+
+        Args:
+            None
+        
+        Returns: 
+            The linear evolution operator of the system.
+        """
+
         k2 = self.calc_k2()
         return - k2 * (self.r + (1 - k2)**2*(4/3-k2)**2)
 
     def calc_stress_tensor_microscopic(self):
-        """
-        Calculates the microscopic stress of the phase-field crystal.
+        """Calculates the microscopic stress of the phase-field crystal.
+
+        Args:
+            None
+
+        Returns:
+            The microscopic stress of the phase-field crystal.
         """
         stress = np.zeros((6,self.xRes,self.yRes,self.zRes))
         
@@ -137,6 +185,14 @@ class PhaseFieldCrystal3DFaceCenteredCubic(PhaseFieldCrystal):
         return stress
 
     def calc_stress_divergence_f(self, field_f = None):
+        """Calculates the divergence of the stress tensor in Fourier space.
+
+        Args:
[…]
