### 2024-04-25 c07595a
Fixed doc for pfc 3d sc
