### 2024-04-25 vidarsko/ComFiT#9 @vidarsko
Thanks for raising the issue. We have updated the Doc strings in all the classes and also fixed the error in the documentation of the BoseEinsteinCondensate class.
