### 2024-04-25 #6599 @vidarsko
> @vidarsko Please have a look at [vidarsko/ComFiT#9](https://github.com/vidarsko/ComFiT/issues/9)

We have addressed and closed the issue. Feel free to raise another issue if you find other problems and or are not satisfied with the resolution.
