### 2024-04-25 #6599 @vidarsko
Note: I have added the custom domain https://comfitlib.com/ to the repository and changed the links in the paper and relevant sections of the code repository. Hope that is okay.
