### 2024-04-26 d5b8bc0
Some striped simulations
