+++ development/240426vs_stripe_demod3.py
@@ -1,0 +1,110 @@
+import os
+import sys
+sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+import comfit as cf
+import numpy as np
+import matplotlib.pyplot as plt
+import scipy as sp
+
+N=16
+T=0.1
+
+pfc = cf.PhaseFieldCrystal2DTriangular(60,23)
+eta = np.exp(1j*pfc.calc_angle_field_vortex_dipole(dipole_position=[0.33*pfc.xmax,pfc.ymax/2], dipole_vector=[pfc.xmax/4,0]))
+theta=0.1
+q0 = [np.cos(theta),np.sin(theta)]
+psi1 = np.real(eta*np.exp(1j*(q0[0]*pfc.x + q0[1]*pfc.y)))
+theta=-0.1
+q0 = [np.cos(theta),np.sin(theta)]
+eta = eta*np.exp(1j*pfc.calc_angle_field_vortex_dipole(dipole_position=[0.66*pfc.xmax,pfc.ymax/2], dipole_vector=[-pfc.xmax/4,0]))
+psi2 = np.real(eta*np.exp(1j*(q0[0]*pfc.x + q0[1]*pfc.y)))
+pfc.psi = psi1
+pfc.psi[pfc.xRes//2:,:] = psi2[pfc.xRes//2:,:]
+pfc.psi_f = np.fft.fft2(pfc.psi)
+pfc.evolve_PFC(500)
+# # pfc.plot_field(pfc.psi)
+# # plt.show()
+
+# pfc = cf.PhaseFieldCrystal2DTriangular(20,13)
+# eta = np.exp(1j*pfc.calc_angle_field_vortex_dipole())
+# # q0 = [1,0]
+# # q0=[np.sqrt(2)/2,np.sqrt(2)/2]
+# # q0 = [0,1]
+# q0 = [np.sqrt(3)/2,1/2]
+# theta=np.pi/N/2
+# q0 = [np.cos(theta),np.sin(theta)]
+# pfc.psi = np.real(eta*np.exp(1j*(q0[0]*pfc.x + q0[1]*pfc.y)))
+
+eta0 = (np.max(pfc.psi)-np.min(pfc.psi))/2
+print(eta0)
+
+
+
+
+def demodulate(pfc,q):
+    return sp.fft.ifftn(sp.fft.fftn(pfc.psi*np.exp(-1j*q[0]*pfc.x - 1j*q[1]*pfc.y))*pfc.calc_Gaussian_filter_f())
+
+
+def orientation(pfc,T):
+    qs = [[np.cos(theta),np.sin(theta)] for theta in np.linspace(0,np.pi-(np.pi)/N,N)]
+    Z = 0
+    p = np.zeros((len(qs),pfc.xRes,pfc.yRes))
+    eta = np.zeros((len(qs),pfc.xRes,pfc.yRes),dtype=complex)
+    for q,n in zip(qs,range(len(qs))):
+        eta[n] = sp.fft.ifftn(sp.fft.fftn(pfc.psi*np.exp(-1j*q[0]*pfc.x - 1j*q[1]*pfc.y))*pfc.calc_Gaussian_filter_f())
+        p[n] = np.exp(abs(eta[n])/T)
+        Z += p[n]
+    
+    Q = np.zeros((3,pfc.xRes,pfc.yRes))
+    delta_q = np.zeros((2,pfc.xRes,pfc.yRes))
+
+    for q,n in zip(qs,range(len(qs))):
+        prob = p[n]/Z
+        eta_f = sp.fft.fftn(eta[n])
[…]

[…]
