### 2024-04-26 #6599 @flokno
@vidarsko regarding the summary in the paper (`Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?`):

For my taste it is a bit too general for a non-expert in PDEs. Maybe you could name one or several widely-known PDEs and physical areas as examples right in the introduction (e.g. Navier-Stokes for fluid dynamcis?)
