### 2024-04-29 d0a10cc
added instructions fro running tests
