### 2024-05-02 87bab2a
Create 240502vs_stripe_demod.py
