### 2024-05-09 #6599 @sharanroongta
Hi @vidarsko, [Here](https://colab.research.google.com/github/vidarsko/ComFiT/blob/main/tutorial/base_system_make_your_own_model.ipynb#scrollTo=YM-ueQmXai8Z) you have mentioned about two examples - one for scalar order parameter, and the other for a vector. I don't see any example for vector order parameter. Am I missing  something?
