### 2024-05-10 #6599 @vidarsko
> Hi @vidarsko, [Here](https://colab.research.google.com/github/vidarsko/ComFiT/blob/main/tutorial/base_system_make_your_own_model.ipynb#scrollTo=YM-ueQmXai8Z) you have mentioned about two examples - one for scalar order parameter, and the other for a vector. I don't see any example for vector order parameter. Am I missing something?

Thanks for pointing this out! I have extended the tutorial to include an example with a vector-valued order parameter.
