### 2024-05-11 4a8c501
Added python tests to ci
