+++ comfit/core/base_system_plot.py
@@ -22,7 +22,7 @@ class BaseSystemPlot:
         elif self.plot_lib == 'plotly':
             go.show() #or fig.show?
 
-    def plot_tool_set_axis_properties(self, **kwargs):
+    def plot_tool_set_axis_properties_matplotlib(self, **kwargs):
         """Sets the properties of the axis for a plot.
         
         Args:
@@ -165,6 +165,50 @@ class BaseSystemPlot:
 
         return ax
 
+    def plot_tool_surface_matplotlib(self, **kwargs):
+        """Plots the surface of the given field.
+
+        Args:
+            **kwargs: Keyword arguments for the plot.
+        
+        Returns:
+            The axes containing the plot.
+        """
+        field = kwargs['field']
+        value = kwargs['value']
+        ax = kwargs.get('ax', plt.gca())
+        alpha = kwargs.get('alpha', 0.5)
+        color = kwargs.get('color', 'b')
+
+        verts, faces, _, _ = marching_cubes(field, value)
+        x = (self.xmin+verts[:, 0]*self.dx)/self.a0
+        y = (self.ymin+verts[:, 1]*self.dy)/self.a0
+        z = (self.zmin+verts[:, 2]*self.dz)/self.a0
+        ax.plot_trisurf(x, y, faces, z, alpha=alpha, color=color)
+    
+    def plot_tool_set_axis_properties_plotly(self, **kwargs):
+        """Sets the properties of the axis for a plot with the plotly library.
+
+        Args:
+            kwargs: keyword arguments for the axis properties
+        
+        Returns:
+            The figure containing the plot.
+        """
+        pass
+
+    def plot_surface_plotly(self, field, **kwargs):
+        """Plots the surface of the given field.
+
+        Args:
+            field (array-like): The field to be plotted.
+            **kwargs: Keyword arguments for the plot.
+        
+        Returns:
+            The figure containing the plot.
+        """
+        pass
+
     def plot_tool_extend_field(self,field):
         """Extends the field in case not a complete array is given. 
 
@@ -207,27 +251,6 @@ class BaseSystemPlot:
 
         return field
 
-    def plot_tool_surface(self, **kwargs):
-        """Plots the surface of the given field.
-
-        Args:
-            **kwargs: Keyword arguments for the plot.
-        
-        Returns:
-            The axes containing the plot.
-        """
-        field = kwargs['field']
-        value = kwargs['value']
-        ax = kwargs.get('ax', plt.gca())
-        alpha = kwargs.get('alpha', 0.5)
-        color = kwargs.get('color', 'b')
-
-        verts, faces, _, _ = marching_cubes(field, value)
-        x = (self.xmin+verts[:, 0]*self.dx)/self.a0
-        y = (self.ymin+verts[:, 1]*self.dy)/self.a0
-        z = (self.zmin+verts[:, 2]*self.dz)/self.a0
-        ax.plot_trisurf(x, y, faces, z, alpha=alpha, color=color)
-    
     def plot_field(self, field, **kwargs):
         """Plots the given (real) field.
         
@@ -238,7 +261,10 @@ class BaseSystemPlot:
                 for a full list of keyword arguments.
         
         Returns:
-            The axes containing the plot (matplotlib.axes.Axes).
+            If self.plot_lib == 'matplotlib':
+                The axes containing the plot (matplotlib.axes.Axes).
+            If self.plot_lib == 'plotly':
+                The figure containing the plot.
         """
 
         if field.dtype == bool:
@@ -250,10 +276,14 @@ class BaseSystemPlot:
             print('Max imaginary part: ', np.max(np.imag(field)))
             field = np.real(field)
 
[…]
