+++ comfit/core/base_system_plot.py
@@ -22,7 +22,7 @@ class BaseSystemPlot:
         elif self.plot_lib == 'plotly':
             go.show() #or fig.show?
 
-    def plot_tool_set_axis_properties_matplotlib(self, **kwargs):
+    def plot_tool_set_axis_properties(self, **kwargs):
         """Sets the properties of the axis for a plot.
         
         Args:
@@ -32,28 +32,29 @@ class BaseSystemPlot:
             The axis object with the properties set.
         """
 
-        # Check if an axis object is provided
-        if 'ax' in kwargs:
-            ax = kwargs['ax']
+        plot_lib = kwargs.get('plot_lib', self.plot_lib)
+        
+        if plot_lib == 'matplotlib':
+            ax = kwargs.get('ax', plt.gca())
+        elif plot_lib == 'plotly':
+            fig = kwargs.get('fig', go.Figure())
 
-        # Set the xlabel
-        if 'xlabel' in kwargs:
-            ax.set_xlabel(kwargs['xlabel'])
-        else:
-            ax.set_xlabel('$x/a_0$')
+        xlabel = kwargs.get('xlabel', 'x/a₀')
 
         if 'title' in kwargs:
                 ax.set_title(kwargs['title'])
 
         #TODO: Check if I should instead include a fig object somewhere (Vidar 05.02.24)
-        if 'suptitle' in kwargs:
-            plt.suptitle(kwargs['suptitle'])
-
-        # Set the grid
-        if 'grid' in kwargs:
-            ax.grid(kwargs['grid'])
-        else:
-            ax.grid(True)
+        suptitle = kwargs.get('suptitle', None)
+        if suptitle is not None:
+            if plot_lib == 'matplotlib':
+                ax.get_figure().suptitle(suptitle)
+            elif plot_lib == 'plotly':
+                fig.update_layout(title_text=suptitle)
+        
+        grid = kwargs.get('grid', True)
+        if plot_lib == 'matplotlib':
+            ax.grid(grid)
 
         xlim = [self.xmin/self.a0, (self.xmax-self.dx)/self.a0]
 
@@ -89,81 +90,150 @@ class BaseSystemPlot:
             zlim = np.array(kwargs['zlim'])/self.a0 if self.dim > 2 else kwargs['zlim']
 
         # Custom x-ticks
-        if 'xticks' in kwargs:
-            ax.set_xticks(kwargs['xticks'])
-        if 'xticklabels' in kwargs:
-            ax.set_xticklabels(kwargs['xticklabels'])
-
-        # Custom y-ticks
-        if 'yticks' in kwargs:
-            ax.set_yticks(kwargs['yticks'])
-        if 'yticklabels' in kwargs:
-            ax.set_yticklabels(kwargs['yticklabels'])
+        xticks = kwargs.get('xticks', None)
+        if xticks is not None:
+            if plot_lib == 'matplotlib':
+                ax.set_xticks(xticks)
         
-        # Custom z-ticks
-        if 'zticks' in kwargs:
-            ax.set_zticks(kwargs['zticks'])
-        if 'zticklabels' in kwargs:
-            ax.set_zticklabels(kwargs['zticklabels'])
+        xticklabels = kwargs.get('xticklabels', None)
+        if xticklabels is not None:
+            if plot_lib == 'matplotlib':
+                ax.set_xticklabels(xticklabels)
+            elif plot_lib == 'plotly':
+                fig.update_layout(xaxis=dict(tickvals=xticks, ticktext=xticklabels))
+
+        yticks = kwargs.get('yticks', None)
+        if yticks is not None:
+            if plot_lib == 'matplotlib':
+                ax.set_yticks(yticks)
+            elif plot_lib == 'plotly':
[…]
