### 2024-05-16 vidarsko/ComFiT#13 @flokno
Error in `base_system_make_your_own_model.ipynb`
Hi,

I am running the `base_system_make_your_own_model` tutorial on my own laptop and get an error when trying to create the gif in the cell

```python
# Initiating a system with positive r value
ls = LandauSystem(2, 0.5)

# Setting an initial condition with both positive and negative values
ls.psi = np.random.rand(ls.xRes, ls.yRes) - 0.5
ls.psi_f = sp.fft.fftn(ls.psi)

# Make animation
for n in range(100):
    ls.evolve(2)
    plt.clf()  # Clearing current figure before plotting new one
    ls.plot_field(ls.psi)
    cf.tool_save_plot(n)
cf.tool_make_animation_gif(n, name="evolution_positive_r")
```

The error reads (a bit annoying output from vscode):

```bash
Cell In[8], [line 14](vscode-notebook-cell:?execution_count=8&line=14)
     [12](vscode-notebook-cell:?execution_count=8&line=12)     ls.plot_field(ls.psi)
     [13](vscode-notebook-cell:?execution_count=8&line=13)     cf.tool_save_plot(n)
---> [14](vscode-notebook-cell:?execution_count=8&line=14) cf.tool_make_animation_gif(n, name="evolution_positive_r")

File ~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/comfit/tools/tool_animation.py:90, in tool_make_animation_gif(counter, name, fps)
     [88](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/comfit/tools/tool_animation.py:88) for image_file in image_files:
     [89](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/comfit/tools/tool_animation.py:89)     img.append(imageio.imread(image_file))
---> [90](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/comfit/tools/tool_animation.py:90) imageio.mimsave(name, img, fps=fps, loop=0)
     [92](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/comfit/tools/tool_animation.py:92) # Delete the png files
     [93](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/comfit/tools/tool_animation.py:93) for file in image_files:

File ~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/imageio/v2.py:495, in mimwrite(uri, ims, format, **kwargs)
    [493](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/imageio/v2.py:493) imopen_args["legacy_mode"] = True
    [494](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/imageio/v2.py:494) with imopen(uri, "wI", **imopen_args) as file:
--> [495](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/imageio/v2.py:495)     return file.write(ims, is_batch=True, **kwargs)

File ~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/imageio/plugins/pillow.py:417, in PillowPlugin.write(self, ndimage, mode, format, is_batch, **kwargs)
    [414](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/imageio/plugins/pillow.py:414)     kwargs["duration"] = 1000 * 1 / kwargs.get("fps")
    [416](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/imageio/plugins/pillow.py:416) if isinstance(ndimage, list):
--> [417](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/imageio/plugins/pillow.py:417)     ndimage = np.stack(ndimage, axis=0)
    [418](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/imageio/plugins/pillow.py:418)     is_batch = True
    [419](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/imageio/plugins/pillow.py:419) else:
...
--> [449](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/numpy/core/shape_base.py:449)     raise ValueError('all input arrays must have the same shape')
    [451](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/numpy/core/shape_base.py:451) result_ndim = arrays[0].ndim + 1
    [452](https://file+.vscode-resource.vscode-cdn.net/Users/flokno/working/projects/peer_review/JOSS/6599_comfit/ComFiT/tutorial/~/working/projects/peer_review/JOSS/6599_comfit/ComFiT/venv/lib/python3.9/site-packages/numpy/core/shape_base.py:452) axis = normalize_axis_index(axis, result_ndim)

ValueError: all input arrays must have the same shape

```

The error apparently has something to do with `pillow`, of which I have version 10.3.0 installed.

I attach a list of my library versions for reference

[packages.txt](https://github.com/vidarsko/ComFiT/files/15336695/packages.txt)
