### 2024-05-16 vidarsko/ComFiT#13 @flokno
edit: it seems to work in the colab notebook. will check my local setup
