### 2024-05-16 vidarsko/ComFiT#13 @flokno
It seems that my images have slightly different sizes when saved locally, which then produces the error.

I changed `comfit.tools.tool_animation.tool_make_animation` to

```python
def tool_make_animation_gif(counter, name=None, fps=24):
    """
    Creates an animation from a series of plot images and saves it as a GIF file.

    Input:
        - counter (int): The number of plot images to include in the animation.
        - name (str, optional): The filename for the output video. Defaults to today's date followed by ' - output_video.mp4'.
        - fps (int, optional): The frames per second for the video. Defaults to 24.
    """

    if name is None:
        name = datetime.now().strftime("%y%m%d_%H%M") + ' - output_animation.gif'
    else:
        name = datetime.now().strftime("%y%m%d_%H%M") + ' - ' + name + '.gif'

    image_files = [f'plot_{counter}.png' for counter in range(counter)]
    images = []
    for image_file in image_files:
        print(image_file)
        image = imageio.imread(image_file)
        print(image.shape)
        images.append(image)
    imageio.mimsave(name, images, fps=fps, loop=0)

    # Delete the png files
    for file in image_files:
        os.remove(file)
```

then I get the output

```
plot_0.png
(435, 525, 4)
plot_1.png
(435, 525, 4)
plot_2.png
(435, 533, 4)
plot_3.png
(435, 533, 4)
plot_4.png
(435, 542, 4)
plot_5.png
(435, 542, 4)
...
```

i.e., the size of the images has changed
