### 2024-05-17 0328e01
Changed save_plot_function to default resolution
