+++ comfit/tools/tool_animation.py
@@ -1,12 +1,10 @@
 """
-
 This module provides tools for saving Matplotlib plots as img and creating animations from these img using MoviePy.
 
 
 Functions:
 
 - tool_save_plot: Saves a Matplotlib plot as an image file.
-
 - tool_make_animation: Creates an animation from a sequence of image files and saves it as a video file.
 """
 
@@ -19,18 +17,21 @@ from datetime import datetime
 import imageio
 
 
-def tool_save_plot(counter):
+def tool_save_plot(counter, image_size=(5,5), dpi=100):
     """
 
     Saves the current Matplotlib plot as a PNG file.
 
-
     Args:
-
-    - counter (int): A unique identifier for the plot image file.
+        counter (int): A unique identifier for the plot image file.
+        image_size (tuple, optional): The size of the image in inches. Defaults to (5,5).
+        dpi (int, optional): The resolution of the image in dots per inch. Defaults to 100.
+    
+    Returns: 
+        None, saves the plot as a PNG file.
     """
-
-    plt.savefig(f'plot_{counter}.png')
+    plt.gcf().set_size_inches(image_size)
+    plt.savefig(f'plot_{counter}.png',dpi=dpi)
 
 
 def tool_make_animation_movie(counter, name=None, fps=24):
@@ -40,12 +41,12 @@ def tool_make_animation_movie(counter, name=None, fps=24):
 
 
     Args:
-
-    - counter (int): The number of plot img to include in the animation.
-
-    - name (str, optional): The filename for the output video. Defaults to today's date followed by ' - output_video.mp4'.
-
-    - fps (int, optional): The frames per second for the video. Defaults to 24.
+        counter (int): The number of plot img to include in the animation.
+        name (str, optional): The filename for the output video. Defaults to today's date followed by ' - output_video.mp4'.
+        fps (int, optional): The frames per second for the video. Defaults to 24.
+    
+    Returns:
+        None, saves the video file.
     """
     
     if name is None:
@@ -72,9 +73,12 @@ def tool_make_animation_gif(counter, name=None, fps=24):
     Creates an animation from a series of plot img and saves it as a GIF file.
 
     Args:
-        - counter (int): The number of plot img to include in the animation.
-        - name (str, optional): The filename for the output video. Defaults to today's date followed by ' - output_video.mp4'.
-        - fps (int, optional): The frames per second for the video. Defaults to 24.
+        counter (int): The number of plot img to include in the animation.
+        name (str, optional): The filename for the output video. Defaults to today's date followed by ' - output_video.mp4'.
+        fps (int, optional): The frames per second for the video. Defaults to 24.
+
+    Returns:
+        None, saves the GIF file.
     """
 
     if name is None:
