### 2024-05-17 vidarsko/ComFiT#13 @vidarsko
Thank you for raising the issue. It seems that the cf.tool_save_plot function might save figures in different formats depending on a users specific setup. I have changed the tool_save_plot-function to take as optional inputs `image_size_inches=(6,5)' and `dpi=100' (dots per inch). If not provided explicitly, this should set the resolution of the plots to 600x500 which will hopefully fix the issue you are having. I tried it on my computer setup and it produced a .gif with the correct dimensions.

Can you pull the main branch and run your script again and see if it fixes your issue?
