### 2024-05-17 af77a28
Changed sentence about Maxwells equations in paper
