### 2024-05-17 #6599 @vidarsko
> > > @vidarsko regarding the summary in the paper (`Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?`):
> > > For my taste it is a bit too general for a non-expert in PDEs. Maybe you could name one or several widely-known PDEs and physical areas as examples right in the introduction (e.g. Navier-Stokes for fluid dynamcis?)
> >
> >
> > Thank you for your comment, @flokno. I have elaborated a little in the Statement of need to include some examples of famous PDEs.
>
> Thank you @vidarsko . I find "_Maxwell's equations_ which dictate how electromagnetic fields encode those well-crafted cat memes" a little bit too colloquial and confusing, but that's just a comment.
>
> I have found some problems when running the tutorials locally and not in the cloud, please see [vidarsko/ComFiT#13](https://github.com/vidarsko/ComFiT/issues/13)

Thank you for your thorough review, @flokno. I suggested a solution to your raised issue. Have a look and see if it is to your satisfaction.

I have also revised the sentece which, I agree, was perhaps a bit too colloquial.
