### 2024-05-17 vidarsko/ComFiT#13 @vidarsko
> It does not work:
>
>
>
> a) the wrong variable name is used in https://github.com/vidarsko/ComFiT/blob/main/comfit/tools/tool_animation.py#L33 (`image_size` vs. `image_size_inches`).
>
>
>
> b) If I fix that, I still get the same error as before. I suspect that the images get cropped when the colorbar labels get too many digits. I attach the first 5 pngs I get below. You can see that they have different widths when the number of digits increases
>
>
>
> ![plot_1](https://github.com/vidarsko/ComFiT/assets/28858706/b282da62-74ed-4c52-a4be-3789dff7ba82)
>
> ![plot_2](https://github.com/vidarsko/ComFiT/assets/28858706/a057034c-e0af-425e-9026-33295931ecd6)
>
> ![plot_3](https://github.com/vidarsko/ComFiT/assets/28858706/d9a375c6-ec97-4a3f-a2fc-af7c9c33f6a8)
>
> ![plot_4](https://github.com/vidarsko/ComFiT/assets/28858706/2e7aaa15-881e-4aa5-87ce-bac84b7f7bd1)
>
> ![plot_5](https://github.com/vidarsko/ComFiT/assets/28858706/0861e023-db45-480a-8a4c-d3acfd28c7aa)
>
>
>
>

Strange! It seems your setup with matplotlb ignores the custom size that I tried to impose. Can you try to vary image_size_inches and dpi and see if you get any changes?
