### 2024-05-17 vidarsko/ComFiT#13 @flokno
I tried different `image_size_inches`, no effect.

I could make it work by forcing the `colorbar` not to push the image boundaries, e.g., like this

```python
from matplotlib.ticker import FuncFormatter

...

            if colorbar:
                _fmt = lambda x, pos: '{:15f}'.format(x)
                cbar = plt.colorbar(pcm, ax=ax, format=FuncFormatter(_fmt), fraction=0.046, pad=0.04)
```

Now, I would recommend to find a more robust solution to create the GIFs than relying on matplotlib to create images of exact same pixel count. This seems to be an open issue in matplotlib, see e.g. https://github.com/matplotlib/matplotlib/issues/8543
