### 2024-05-19 f50a7f2
Proposed solution to plot making function
