+++ development/240517vs_testing_plot_changes.py
@@ -5,7 +5,7 @@ import comfit as cf
 import numpy as np
 import matplotlib.pyplot as plt
 import scipy as sp
-
+import numpy as np
 
 class LandauSystem(cf.BaseSystem):
     
@@ -50,9 +50,9 @@ ls.psi = np.random.rand(ls.xRes,ls.yRes)-0.5
 ls.psi_f = sp.fft.fftn(ls.psi)
 
 # Make animation
-for n in range(100):
+for n in range(10):
     ls.evolve(2)
     plt.clf() #Clearing current figure before plotting new one
     ls.plot_field(ls.psi)
-    cf.tool_save_plot(n)
+    cf.tool_save_plot(n,image_size_inches=(6,5), dpi=100+np.random.randint(2))
 cf.tool_make_animation_gif(n,name='evolution_negative_r')
