+++ comfit/core/base_system_plot.py
@@ -7,6 +7,7 @@ from comfit.tools.tool_create_orthonormal_triad import tool_create_orthonormal_t
 import matplotlib.pyplot as plt
 import matplotlib.colors as mcolors
 import matplotlib.tri as mtri
+import matplotlib.cm as cm
 
 import plotly.graph_objects as go
 
@@ -300,6 +301,37 @@ class BaseSystemPlot:
 
         return field
 
+    def plot_tool_plotly_add_angle_colorbar3D(self,fig):
+        """Adds a colorbar to a 3D plot with the angle colormap.
+
+        Args:
+            fig: The figure to which the colorbar is added.
+        
+        Returns:
+            The figure with the colorbar added.
+        """
+        # Generate the custom colormap
+        custom_colormap = tool_colormap_angle(pyplot=True)
+
+        fig.add_trace(go.Isosurface(
+            x=[self.xmin/self.a0,self.xmax/self.a0], 
+            y=[self.ymin/self.a0,self.ymax/self.a0], 
+            z=[self.zmin/self.a0, self.zmax/self.a0],
+            value=[0,0],
+            cmin = -np.pi,
+            cmax = np.pi,
+            opacity=0,
+            colorscale=custom_colormap,
+            colorbar=dict(
+                tickvals=[-np.pi, -2*np.pi/3, -np.pi/3, 0, np.pi/3, 2*np.pi/3, np.pi],
+                ticktext=['-π', '-2π/3', '-π/3', '0', 'π/3', '2π/3', 'π']
+            ),
+            showscale=True
+        ))
+
+        return fig
+
+
     def plot_field(self, field, **kwargs):
         """Plots the given (real) field.
         
@@ -336,6 +368,7 @@ class BaseSystemPlot:
 
         # Kewyord arguments
         colorbar = kwargs.get('colorbar', True)
+        axis_equal = kwargs.get('axis_equal',True)
 
         # Extend the field if not a complete array is given
         field = self.plot_tool_extend_field(field)
@@ -357,90 +390,114 @@ class BaseSystemPlot:
             # Keyword arguments particular to the 2D case
             kwargs['grid'] = kwargs.get('grid', False)
 
-            if ax == None:
-                fig.clf()
-                ax = fig.add_subplot(111)
-            
-            # Set the colormap
-            colormap = kwargs.get('colormap', 'viridis')
-
-            if colormap == 'bluewhitered':
-                colormap = tool_colormap_bluewhitered()
-
-            elif colormap == 'sunburst':
-                colormap = tool_colormap_sunburst()
-            else:
-                colormap = plt.get_cmap(colormap)
-
-            # Value limits symmetric
-            vlim_symmetric = kwargs.get('vlim_symmetric', False)
-
-            X, Y = np.meshgrid(self.x, self.y, indexing='ij')
-
-            pcm = ax.pcolormesh(X / self.a0, Y / self.a0, field, shading='gouraud', cmap=colormap)
-
-            xlim = [self.xmin, self.xmax-self.dx]
-            ylim = [self.ymin, self.ymax-self.dy]
-
-            limits_provided = False
-            if 'xlim' in kwargs:
-                xlim = kwargs['xlim']
-                limits_provided = True
-            else:
-                if 'xmin' in kwargs:
-                    xlim[0] = kwargs['xmin']
-                    limits_provided = True
+            if self.plot_lib == 'matplotlib':
+                if ax == None:
+                    fig.clf()
+                    ax = fig.add_subplot(111)
                 
-                if 'xmax' in kwargs:
-                    xlim[1] = kwargs['xmax']
-                    limits_provided = True
+                # Set the colormap
+                colormap = kwargs.get('colormap', 'viridis')
 
[…]
