### 2024-05-20 vidarsko/ComFiT#13 @flokno
> I have added another check in the `tool_make_animation_gif`, whereby it checks if all the images are of the same dimensions and crops them if not. As a side note, I realized that since imageio relies on the pillow package, which can do the same job, I was able to remove the dependency on imageio from comfit, which should be a little bit more stable.
>
> I was not able to reproduce the error with the packages you listed, but let me suggest three ways of exploring this issue:
>
>     1. Can you try to run comfit with this latest patch and see if the proposed patch works (even though it is not super elegant)?

I think it works now. I see the warning when the files are cropped, and I get the GIFs 👍

>     2. I see that in the list of dependencies, you have comfit==1.4.2 installed. Is it possible that when you are running the script with `import comfit`, that uses the current version from pypi (which does not have the patch implemented yet)? Could you make a clean virtual environnement with a clone of the github repository and only the required packages and see if it fixes the issue?

I was working with a clean `venv` and commented out the pip install.

> I agree that saving images as pngs and then to export them is not an optimal strategy, and I will think of future ways of making it better. If neither step 1 or 2 works, would you be open to leaving this issue open and still continuing with the submission of the paper, or do you think that this is essential for the package to be published in JOSS?

I think making the GIFs is a very nice feature so it is nice when it actually works. This seems to be the case now!
