### 2024-05-20 #6599 @flokno
@vidarsko I tried the other tutorials and ran into a bunch of errors, see issues. I stopped for now -- please double-check all tutorials and let me know when they are ready. Once all tutorials works, I'm fine with the JOSS paper!

As an optional comment: there are tools to auto-format python files and jupyter notebooks. For example: https://github.com/drillan/jupyter-black

I strongly recommend consistent code formatting complying with [PEP8](https://peps.python.org/pep-0008/) (and possibly tools like `flake8` to check), though this is of course optional
