### 2024-05-24 #6599 @flokno
@jonasron @vidarsko good, thank you! I ticked the last box.

@lucydot I'm done with my review and I think the package is good to go!
