### 2024-05-30 #6599 @lucydot
Excellent work @flokno @sharanroongta @sbacchio; on behalf of the JOSS team, thank you for your expertise and time.
