### 2024-05-30 #6599 @lucydot
@vidarsko, some comments on the paper:

- The sentence spanning lines 58-62 is rather long (and "between" is used twice). I think this could be split into two sentences.
- Acronym `UiO`is undefined.
- I've made several other small corrections to improve readability. Please see PR here, and double check all changes before merging: https://github.com/vidarsko/ComFiT/pull/19

Once you have updated the paper, please run `@editorialbot generate pdf` to double check that the paper compiles and reads as expected.

Nearly there!
