### 2024-05-30 #6599 @lucydot
@vidarsko for the cited doctoral theses, please can you double check if there are DOIs associated. Usually there is an institutional record with an assigned DOI. Thanks.
