+++ JOSS_paper/paper.md
@@ -54,10 +54,11 @@ The project aims to be highly community-driven, continuously improving the code 
 
 ComFiT is a synthesized code base that was originally written in Matlab and used to fuel research into projects exploring the statistical properties and similarities between the dynamics of topological defects between Bose-Einstein condensates and phase-field crystals [@skaugenUnifiedPerspectiveTwodimensional2018] building on the framework for evolving stiff numerical systems described in Ref. [@coxExponentialTimeDifferencing2002] and the methods for tracking defects developed in Refs. [@mazenkoVortexVelocitiesSymmetric1997;@mazenkoVelocityDistributionStrings1999;@anghelutaAnisotropicVelocityStatistics2012].
 The code base was further improved, stabilized and extended in the PhD projects of the paper authors [@ronningTopologicalDefectsFlows2023;@skogvollSymmetryTopologyCrystal2023].
-ComFiT has been used for research into a range of physical systems, including active matter [@ronningSpontaneousFlowsDynamics2023], stirred Bose-Einstein condensates [@ronningClassicalAnalogiesForce2020;@ronningPrecursoryPatternsVortex2023] and phase-field crystals [@skogvollDislocationNucleationPhasefield2021;@skogvollStressOrderedSystems2021;@skogvollPhaseFieldCrystal2022;@skogvollHydrodynamicPhaseField2022].
-In all of these applications, the role of topological defects is central to understanding the critical dynamics and the statistical properties of the systems. For example, in Bose-Einstein condensates the formation of vortices mark the transition to quantum turbulence and in crystals, and the formation of dislocations are central to understanding how materials buckle.
+ComFiT has been used for research into a range of physical systems, including active matter [@ronningSpontaneousFlowsDynamics2023], stirred Bose-Einstein condensates [@ronningClassicalAnalogiesForce2020;@ronningPrecursoryPatternsVortex2023], and phase-field crystals [@skogvollDislocationNucleationPhasefield2021;@skogvollStressOrderedSystems2021;@skogvollPhaseFieldCrystal2022;@skogvollHydrodynamicPhaseField2022].
+In all of these applications, the role of topological defects is central to understanding the critical dynamics and the statistical properties of the systems.
+For example, in Bose-Einstein condensates, the formation of vortices marks the transition to quantum turbulence, whereas in crystals, the formation of dislocations is central to understanding how materials buckle.
 The library is, at the time of publication, being used to research turbulence in 3D active nematics and the dynamics of 3D vortex structures in Bose-Einstein condensates.
-The Phase Field Crystal framework is currently under exploration at the Njord Centre at the University of Oslo for modeling crack propagation in Earth's upper crust and the basic framework is used as a foundation for several projects in the theoretical physics group at Porelab, UiO.
+The phase-field crystal framework is currently under exploration at the Njord Centre at the University of Oslo for modeling crack propagation in Earth's upper crust and the basic framework is used as a foundation for several projects in the theoretical physics group at Porelab at the University of Oslo.
 
 # Appendix
 
