+++ JOSS_paper/paper.md
@@ -52,7 +52,8 @@ The project aims to be highly community-driven, continuously improving the code 
 
 # Research projects 
 
-ComFiT is a synthesized code base that was originally written in Matlab and used to fuel research into projects exploring the statistical properties and similarities between the dynamics of topological defects between Bose-Einstein condensates and phase-field crystals [@skaugenUnifiedPerspectiveTwodimensional2018] building on the framework for evolving stiff numerical systems described in Ref. [@coxExponentialTimeDifferencing2002] and the methods for tracking defects developed in Refs. [@mazenkoVortexVelocitiesSymmetric1997;@mazenkoVelocityDistributionStrings1999;@anghelutaAnisotropicVelocityStatistics2012].
+ComFiT is a synthesized code base that was originally written in Matlab and used to fuel research into projects exploring the statistical properties and similarities between the dynamics of topological defects between Bose-Einstein condensates and phase-field crystals [@skaugenUnifiedPerspectiveTwodimensional2018].
+It was built on the framework for evolving stiff numerical systems described in Ref. [@coxExponentialTimeDifferencing2002] and the methods for tracking defects developed in Refs. [@mazenkoVortexVelocitiesSymmetric1997;@mazenkoVelocityDistributionStrings1999;@anghelutaAnisotropicVelocityStatistics2012].
 The code base was further improved, stabilized and extended in the PhD projects of the paper authors [@ronningTopologicalDefectsFlows2023;@skogvollSymmetryTopologyCrystal2023].
 ComFiT has been used for research into a range of physical systems, including active matter [@ronningSpontaneousFlowsDynamics2023], stirred Bose-Einstein condensates [@ronningClassicalAnalogiesForce2020;@ronningPrecursoryPatternsVortex2023], and phase-field crystals [@skogvollDislocationNucleationPhasefield2021;@skogvollStressOrderedSystems2021;@skogvollPhaseFieldCrystal2022;@skogvollHydrodynamicPhaseField2022].
 In all of these applications, the role of topological defects is central to understanding the critical dynamics and the statistical properties of the systems.
