### 2024-05-30 #6599 @vidarsko
Thank you. I made some minor changes to the pull request before merging, split the sentence for readability and have written out the acronym.

The University of Oslo does not assign DOIs to doctoral theses. I do, however, have urls for where these are hosted on the Universitys research archive: [Skogvoll](https://www.duo.uio.no/handle/10852/102731) [Rønning](https://www.duo.uio.no/handle/10852/104678). Do you want med to add these URLs to the bibliography?
