### 2024-05-30 9c9a3f8
Updated to version 1.5.0 before JOSS publication
