### 2024-05-30 #6599 @vidarsko
Author check list:

Double checked. OK.

Updated to version 1.5.0 which is released both on github and pypi.

doi.org/10.5281/zenodo.11395524

Done. I have also changed the date of the paper to today (30 May 2024).

Checked. Both are MIT.
