+++ comfit/core/base_system_plot.py
@@ -10,6 +10,7 @@ import matplotlib.tri as mtri
 import matplotlib.cm as cm
 
 import plotly.graph_objects as go
+import plotly.figure_factory as ff
 
 from skimage.measure import marching_cubes
 
@@ -1112,9 +1113,10 @@ class BaseSystemPlot:
             print('Max imaginary part: ', np.max(np.imag(vector_field)))
             vector_field = np.real(vector_field)
 
-        # Check if an axis object is provided
-        fig = kwargs.get('fig', plt.gcf())
-        ax = kwargs.get('ax', None)
+        if self.plot_lib == 'matplotlib':
+            # Check if an axis object is provided
+            fig = kwargs.get('fig', plt.gcf())
+            ax = kwargs.get('ax', None)
 
         def add_spacing_2D(X,Y,U,V,spacing):
             X = X[::spacing, ::spacing]
@@ -1135,9 +1137,11 @@ class BaseSystemPlot:
         if self.dim == 1:
             
             if vector_field.shape == (1,self.xRes):
-                if ax == None:
-                    fig.clf()
-                    ax = plt.gcf().add_subplot(111)
+
+                if self.plot_lib == 'matplotlib':
+                    if ax == None:
+                        fig.clf()
+                        ax = plt.gcf().add_subplot(111)
 
                 X, Y = np.meshgrid(self.x, np.array([0]), indexing='ij')
 
@@ -1148,14 +1152,17 @@ class BaseSystemPlot:
 
                 X,Y,U,V = add_spacing_2D(X,Y,U,V,spacing)
 
-                ax.quiver(X/self.a0, Y/self.a0, U, V, color='blue', angles='xy', scale_units='xy', scale=1)
+                if self.plot_lib == 'matplotlib':
+                    ax.quiver(X/self.a0, Y/self.a0, U, V, color='blue', angles='xy', scale_units='xy', scale=1)
                 
                 kwargs['ylim'] = [np.min(vector_field[0]), np.max(vector_field[0])]
 
             elif vector_field.shape == (2,self.xRes):
-                if ax == None:
-                    fig.clf()
-                    ax = fig.add_subplot(111, projection='3d')
+                
+                if self.plot_lib == 'matplotlib':
+                    if ax == None:
+                        fig.clf()
+                        ax = fig.add_subplot(111, projection='3d')
                 
                 X, Y, Z = np.meshgrid(self.x, np.array([0]), np.array([0]), indexing='ij')
 
@@ -1168,7 +1175,8 @@ class BaseSystemPlot:
 
                 X,Y,Z,U,V,W = add_spacing_3D(X,Y,Z,U,V,W,spacing)
 
-                ax.quiver(X/self.a0, Y/self.a0, Z/self.a0, U, V, W, color='blue')
+                if self.plot_lib == 'matplotlib':
+                    ax.quiver(X/self.a0, Y/self.a0, Z/self.a0, U, V, W, color='blue')
 
                 kwargs['ylim'] = [np.min(vector_field[0]), np.max(vector_field[0])]
                 delta_y = kwargs['ylim'][1] - kwargs['ylim'][0]
@@ -1183,10 +1191,11 @@ class BaseSystemPlot:
                     kwargs['zlim'] = [kwargs['zlim'][0] - 0.15*delta_y, kwargs['zlim'][1] + 0.15*delta_y]
 
             elif vector_field.shape == (3,self.xRes):
-
-                if ax == None:
-                    fig.clf()
-                    ax = fig.add_subplot(111, projection='3d')
+                
+                if self.plot_lib == 'matplotlib':
+                    if ax == None:
+                        fig.clf()
+                        ax = fig.add_subplot(111, projection='3d')
                 
[…]
