### 2024-06-04 32e045b
Update ClassBoseEinsteinCondensate.md
