### 2024-06-04 b37bfef
Update ClassBoseEinsteinCondensate.md
