### 2024-06-04 c7a1b23
Update ClassNematicLiquidCrystal.md
