### 2024-06-04 fd9678e
Update ClassNematicLiquidCrystal.md
