### 2024-06-04 e0402d3
Update ClassNematicLiquidCrystal.md
