### 2024-06-04 #6599 @lucydot
Thanks @vidarsko: could you add a DOI for pyclaw citation, I can see one given here: https://epubs.siam.org/doi/abs/10.1137/110856976. I think adding URLs is a good idea, especially as they are associated with a "unique identifier" on the UiO repository.

All else looks good with the paper :)
