### 2024-06-04 #6599 @lucydot
@vidarsko a couple of other points:

- the title associated with the Zenodo archive does not look correct. It is currently `vidarsko/ComFiT: ComFiT v1.5.0`. It should exactly match that of the paper `ComFiT: a Python library for computational field
2 theory with topological defects`. You should be able to update this meta-data without having to create a new release, but let me know if you bump into any problems (the Zenodo interface seems to have changed recently).
- the license is listed as Creative Commons on Zenodo, and MIT on Github. Please update so that they are the same.
