### 2024-06-04 40ffbe2
Added doi to pyclaw and links to phds
