+++ JOSS_paper/paper.bib
@@ -93,7 +93,8 @@
   author = {Skaugen, Audun},
   year = {2018},
   school = {The University of Oslo},
-  keywords = {Phase-field crystal}
+  keywords = {Phase-field crystal}, 
+  howpublished = {\url{https://www.duo.uio.no/handle/10852/66179}}
 }
 
 @article{skaugenVelocityStatisticsNonuniform2016,
@@ -207,7 +208,8 @@
   year = {2023},
   urldate = {2023-09-24},
   langid = {english},
-  annotation = {Accepted: 2023-07-13T11:29:00Z}
+  annotation = {Accepted: 2023-07-13T11:29:00Z}, 
+  howpublished = {\url{https://www.duo.uio.no/handle/10852/102731}}
 }
 
 
@@ -218,7 +220,8 @@
   year = {2023},
   urldate = {2024-03-05},
   langid = {english},
-  annotation = {Accepted: 2023-09-08T13:40:41Z}
+  annotation = {Accepted: 2023-09-08T13:40:41Z}, 
+  howpublished = {\url{https://www.duo.uio.no/handle/10852/104678}}
 }
 
 @article{ronningSpontaneousFlowsDynamics2023,
@@ -284,7 +287,8 @@
   journal = {SIAM Journal on Scientific Computing},
   volume = {34},
   number = {4},
-  pages = {C210--C231}
+  pages = {C210--C231}, 
+  doi = {10.1137/110856976}
 }
 
 @manual{hamFiredrakeUserManual2023,
