### 2024-06-04 6cc7695
Updated phd references so they look the same
