+++ JOSS_paper/paper.bib
@@ -201,7 +201,6 @@
 }
 
 @phdthesis{skogvollSymmetryTopologyCrystal2023,
-  type = {Doctoral Thesis},
   title = {Symmetry, Topology, and Crystal Deformations: A Phase-Field Crystal Approach},
   shorttitle = {Symmetry, Topology, and Crystal Deformations},
   author = {Skogvoll, Vidar},
@@ -214,7 +213,6 @@
 
 
 @phdthesis{ronningTopologicalDefectsFlows2023,
-  type = {Doctoral Thesis},
   title = {Topological {{Defects}} and {{Flows}} in {{BECs}} and {{Active Matter}}},
   author = {R{\o}nning, Jonas},
   year = {2023},
