### 2024-06-04 #6599 @vidarsko
@lucydot Thanks!

I have added links to the PhD theses, changed the name on Zenodo and the changed the license.
