### 2024-06-04 fe716ab
Added some quiver plot code. Not optimal
