### 2024-06-05 1af4c07
Made functionning vector field function inplotly
