+++ comfit/core/base_system_plot.py
@@ -15,6 +15,8 @@ import plotly.colors as pc
 
 from skimage.measure import marching_cubes
 
+import time
+
 class BaseSystemPlot:
     """ Plotting methods for the base system class"""
 
@@ -1092,19 +1094,22 @@ class BaseSystemPlot:
 
             return self.plot_complex_field(complex_field, **kwargs)
 
-    def plot_vector_field(self, vector_field, spacing=5, **kwargs):
+    def plot_vector_field(self, vector_field, **kwargs):
         """Plots a vector field.
 
         Args:
             vector_field (tuple): Tuple containing the x and y components of the vector field.
-            spacing (int, optional): The spacing for the quiver plot. Default is 5.
-            **kwargs: Keyword arguments for the plot, see https://comfitlib.com/Plotting/.
+            **kwargs: 
+                spacing (int, optional): The spacing for the quiver plot. Default is self.xRes//20.
+                Keyword arguments for the plot, see https://comfitlib.com/Plotting/.
 
         Returns:
             Tuple containing
                 - The figure containing the plot.
                 - The axes containing the plot.
         """
+        spacing = kwargs.get('spacing', self.xRes//20)
+
         # Convert the vector field to a numpy array
         vector_field = np.array(vector_field)
         
@@ -1307,77 +1312,13 @@ class BaseSystemPlot:
                         angles='xy', scale_units='xy', scale=1,
                         headwidth=3, headlength=4, headaxislength=3, pivot='middle')
                 elif self.plot_lib == 'plotly':
-                    # fig = ff.create_quiver(X/self.a0, Y/self.a0, U, V, line=dict(width=1))
-                    # fig = go.Figure(data=go.Cone(x=X.flatten()/self.a0,
-                    #             y=Y.flatten()/self.a0,
-                    #             z=np.zeros(X.shape).flatten(),
-                    #             u=U.flatten(),
-                    #             v=V.flatten(),
-                    #             w=np.zeros(X.shape).flatten(), 
-                    #             sizemode="absolute",
-                    #             colorscale='Viridis',
-                    #             sizeref=0.1))
-
-                    # Calculate magnitude and angle of the vectors
-                    # u = U.flatten()
-                    # v = V.flatten()
-                    # magnitude = np.sqrt(u**2 + v**2)
-                    # angle = np.arctan2(v, u)
-                    # x = X.flatten()/self.a0
-                    # y = Y.flatten()/self.a0
-
-                    # Ntheta = 36  # Number of angle bins
-                    # Nm = 1  # Number of magnitude bins
-
-                    # # Discretize angles and magnitudes
-                    # theta_bins = np.linspace(-np.pi, np.pi, Ntheta + 1)
-                    # mag_bins = np.linspace(0, np.max(magnitude), Nm + 1)
-                    
-                    
-
-                    # # Create scatter plot data
-                    # scatter_data = []
-                    # for i in range(Ntheta):
-                    #     for j in range(Nm):
-                    #         # Filter points in the current angle and magnitude bin
-                    #         angle_mask = (angle >= theta_bins[i]) & (angle < theta_bins[i+1])
-                    #         mag_mask = (magnitude >= mag_bins[j]) & (magnitude < mag_bins[j+1])
-                    #         mask = angle_mask & mag_mask
-                            
-                    #         # Get points that satisfy the mask
-                    #         x_bin = x[mask]
-                    #         y_bin = y[mask]
-                    #         u_bin = u[mask]
-                    #         v_bin = v[mask]
-                    #         mag_bin = magnitude[mask]
-                            
-                    #         # Add scatter plot for this bin
[…]
