### 2024-06-11 #6599 @lucydot
Note that the missing DOIs are for PhD theses (where DOIs have not been assigned). In this case, links to institutional repository has been provided instead.
