### 2024-06-11 #6599 @lucydot
@vidarsko I've recommended acceptance - great to work with you, and thank you for your submission to JOSS!
