### 2024-06-11 #6599 @kyleniemeyer
Hi @vidarsko, I made a few small edits to the paper: https://github.com/vidarsko/ComFiT/pull/20. Could you review and merge those?
